+++ clang-tools-extra/clang-tidy/misc/MyReturnCheckCheck.cpp
@@ -24,11 +24,19 @@ void MyReturnCheckCheck::registerMatchers(MatchFinder *Finder) {
 
   
   auto direct_return_nullptr_pattern =
-   functionDecl(has(compoundStmt(has(returnStmt(has(implicitCastExpr(has(cxxNullPtrLiteralExpr())))).bind("direct")))));
+    functionDecl(has(compoundStmt(has(returnStmt(has(implicitCastExpr(has(cxxNullPtrLiteralExpr())))).bind("direct_return_stmt"))))).bind("owner_function");
 
   Finder->addMatcher(direct_return_nullptr_pattern, this);
+
+
+  auto sub_node_return = functionDecl(hasDescendant(returnStmt(has(implicitCastExpr(has(cxxNullPtrLiteralExpr().bind("null_expr"))))))).bind("top_func");    
+  Finder->addMatcher(sub_node_return, this);
+
+
   
 }
+
+  
 
 void MyReturnCheckCheck::check(const MatchFinder::MatchResult &Result) {
 
@@ -37,11 +45,19 @@ void MyReturnCheckCheck::check(const MatchFinder::MatchResult &Result) {
     diag(MatchedIfReturn->getLocation(), "found %0 as if return")
       << MatchedIfReturn;
 
-  const auto *MatchedDirectReturn = Result.Nodes.getNodeAs<ReturnStmt>("direct");
-  if(MatchedDirectReturn)
-    diag(MatchedDirectReturn->getReturnLoc(), "found as direct return statement");
+  const auto *MatchedDirectReturn = Result.Nodes.getNodeAs<ReturnStmt>("direct_return_stmt");
+  const auto *function = Result.Nodes.getNodeAs<FunctionDecl>("owner_function");
   
-  
+  if(MatchedDirectReturn && function)
+    diag(MatchedDirectReturn->getReturnLoc(), "found as direct return statement in function %0")
+    << function->getName()  ;
+
+
+  const auto *func = Result.Nodes.getNodeAs<FunctionDecl>("top_func");
+  const auto *expr = Result.Nodes.getNodeAs<CXXNullPtrLiteralExpr>("null_expr");  
+  if(func && expr){
+    diag(expr->getLocation(),"Matched top level function");
+  }
 }
 
 } // namespace misc